edited so that this use case interactor calls subordinate interactors to perform the tasks of schedudling events. this will primarily be used to send scheduled reminders which is more difficult
